FrameWork Infinity: a-1.0
Alpha-1.0 Release; This is the first official release of the Infinity FrameWork project.

The purpose of this project is to allow the safe creation of what I call an Infinity Website.
These are websites that exist within the entire observable spacetime of host (your website space).

They universe provides us a great deal,. This system, is my way of sharing how I understand the universe of web development.

Please help me support it anyway you can! I have a lacking professional background in this field and have worked very hard for a great portion of my life, harnessing my techniques.